moved grobid trainer args to separate module
